4th iteration

+++ 0x15-file_io/0-read_textfile.c
@@ -10,7 +10,7 @@ ssize_t read_textfile(const char *filename, size_t letters)
 {
 	int fd;
 	char *buffer;
-	ssize_t read_count, write_count;
+	int read_count, write_count;
 
 	if (filename == NULL)
 		return (0);
